Refactor out functions into their own files
